Refactor: Extract `detect::macos` in crate `detect-targets`
